requirements.txt
black==24.8.0
click==8.1.7
filelock==3.15.4
fsspec==2024.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
PySocks==1.7.1
sympy==1.13.2
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
